keep on top when the taskbar is focused and after closing start menu
There is still a bug where, if you open the start menu and then click
the taskbar to close it, we stay behind the taskbar. This is bizarre,
since even using a timer delay of 1 second doesn't fix it (maybe the
taskbar gets into some special "really topmost" mode). But selecting any
other window and then selecting something else fixes it, so it's not a
big deal.