switch back to perMachine installation
This is necessary to use elevation, to install to `C:\Program Files`.

Maybe we should be installing to localappdata or something,
but people seem to recommend against that.
Rainmeter also installs to `C:\Program Files`, even though it adds a
per-user shortcut in `shell:startup`. Not sure how it handles multiple
users trying to install it... But I don't need to support that usecase
right now anyways.