guard against zero-size draw rect
This is what has been causing the periodic "BeginBufferedPaint
failures".

It seems that sometimes (perhaps when the taskbar / explorer.exe is
shut down when we enter sleep), the height of the taskbar area becomes
zero. This would then be persisted as the draw rectangle, causing all
subsequent BeginBufferedPaint calls to fail as a result.