fix drawing over top of window contents
Previously, I was seeing a weird behavior where, if the UI was redrawn
rapidly, e.g. when muting and unmuting the mic in quick succession, the
window contents would be drawn over top of each other (like the
Solitaire win screen).

Seems like this is because the buffered paint buffer may be reused
between renders, so for proper behavior it needs to be cleared every
time. (I suppose this makes sense--for opaque UI, clearing would simply
be a waste of time, since it'll all be drawn over anyways.) But the
buffer seems to only be kept for a second or so, so the old behavior
of only redrawing every 5 seconds was accidentally mitigating this bug.